Replaced many g_return* with g_assert.
	* src/favorites.c: (add_single_to_blacklist):
	* src/interface.c: (menu_item_select_cb):
	* src/load-graph.c: (load_graph_unalloc), (load_graph_alloc):
	* src/memmaps.c: (compare_memmaps), (memmaps_timer),
	(create_single_memmaps_dialog):
	* src/openfiles.c: (update_openfiles_dialog), (openfiles_timer),
	(create_single_openfiles_dialog):
	* src/procactions.c: (renice_single_process),
	(kill_single_process):
	* src/procman.c: (color_changed_cb), (procman_get_tree_state),
	(procman_save_tree_state), (procman_save_config), (main):
	* src/proctable.c: (sort_ints), (proctable_free_info),
	(remove_info_from_tree):

	Replaced many g_return* with g_assert.

+++ src/procactions.c
@@ -44,7 +44,9 @@ renice_single_process (GtkTreeModel *model, GtkTreePath *path, GtkTreeIter *iter
 	GtkWidget *dialog;
 
 	gtk_tree_model_get (model, iter, COL_POINTER, &info, -1);
-	g_return_if_fail (info);
+
+	if (!info)
+		return;
 
 	error = setpriority (PRIO_PROCESS, info->pid, args->nice_value);
 
@@ -111,7 +113,9 @@ kill_single_process (GtkTreeModel *model, GtkTreePath *path, GtkTreeIter *iter, 
 	GtkWidget *dialog;
 
 	gtk_tree_model_get (model, iter, COL_POINTER, &info, -1);
-	g_return_if_fail (info);
+
+	if (!info)
+		return;
 
 	error = kill (info->pid, args->signal);
